Removed Old Dashboards and fixed Modal Bug
